Refactor: Elevate L7 Address Domain Knowledge (Phase 2.6 - Part 1)
This commit decouples OSI Layer 3 and Layer 7 addressing to enforce the new architectural boundary. Domain-specific knowledge (e.g., DEV_TYPE_MAP, logical device lookup) has been extracted from the transport layer and elevated to the domain layer.

- Extracted domain addressing logic into `ramses_rf/address.py`.
- Refactored `ramses_tx/address.py` into a pure L2/L3 MAC validator, functioning temporarily as a backward-compatible proxy to satisfy legacy unmigrated L3 builders (scheduled for deletion in Phase 5).
- Updated downstream L7 routing components (`device/`, `system/`, `messages/`) to strictly consume from the new `ramses_rf` boundary.
- Updated the test suite (e.g., `test_hgi_behaviors.py`) to correctly monkeypatch the newly elevated strict-checking flags.

Ref: #639 (OSI Decoupling - Phase 2.6, Part 1)
